use unsigned short better intead of mixing with ints to prevent compiler warnings

+++ ares/ares_query.c
@@ -1,4 +1,4 @@
-/* $Id: ares_query.c,v 1.18 2008-11-20 07:41:26 bagder Exp $ */
+/* $Id: ares_query.c,v 1.19 2008-11-20 07:50:48 bagder Exp $ */
 
 /* Copyright 1998 by the Massachusetts Institute of Technology.
  *
@@ -67,7 +67,7 @@ void ares__rc4(rc4_key* key, unsigned char *buffer_ptr, int buffer_len)
   key->y = y;
 }
 
-static struct query* find_query_by_id(ares_channel channel, int id)
+static struct query* find_query_by_id(ares_channel channel, unsigned short id)
 {
   unsigned short qid;
   struct list_node* list_head;
@@ -94,11 +94,11 @@ static struct query* find_query_by_id(ares_channel channel, int id)
 */
 static unsigned short generate_unique_id(ares_channel channel)
 {
-  short id;
+  unsigned short id;
 
   do {
-	id = ares__generate_new_id(&channel->id_key);
-  } while (find_query_by_id(channel,id));
+    id = ares__generate_new_id(&channel->id_key);
+  } while (find_query_by_id(channel, id));
 
   return (unsigned short)id;
 }
